Adding --help-json to get JSON encoded help message
This adds the `--help-json` flag which prints the help information
encoded in a JSON object. This enables template engines to render
the help information in other formats such as markdown.
